+++ stats.json
@@ -1,0 +1,1 @@
+{"hash":"d1f97b8a4e922c1c2d66","version":"5.91.0","time":75,"builtAt":1723615790076,"publicPath":"auto","outputPath":"/Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/dist","assetsByChunkName":{"main":["main.js"]},"assets":[{"type":"asset","name":"main.js","size":0,"emitted":false,"comparedForEmit":false,"cached":true,"info":{"javascriptModule":false,"minimized":true},"chunkNames":["main"],"chunkIdHints":[],"auxiliaryChunkNames":[],"auxiliaryChunkIdHints":[],"related":{},"chunks":[792],"auxiliaryChunks":[],"isOverSizeLimit":false}],"chunks":[{"rendered":true,"initial":true,"entry":true,"recorded":false,"size":0,"sizes":{},"names":["main"],"idHints":[],"runtime":["main"],"files":["main.js"],"auxiliaryFiles":[],"hash":"b21f67f9136653234533","childrenByOrder":{},"id":792,"siblings":[],"parents":[],"children":[],"modules":[],"origins":[{"module":"","moduleIdentifier":"","moduleName":"","loc":"main","request":"./src"}]}],"modules":[],"entrypoints":{"main":{"name":"main","chunks":[792],"assets":[{"name":"main.js"}],"filteredAssets":0,"assetsSize":null,"auxiliaryAssets":[],"filteredAuxiliaryAssets":0,"auxiliaryAssetsSize":0,"children":{},"childAssets":{},"isOverSizeLimit":false}},"namedChunkGroups":{"main":{"name":"main","chunks":[792],"assets":[{"name":"main.js"}],"filteredAssets":0,"assetsSize":null,"auxiliaryAssets":[],"filteredAuxiliaryAssets":0,"auxiliaryAssetsSize":0,"children":{},"childAssets":{},"isOverSizeLimit":false}},"errors":[{"loc":"main","message":"Module not found: Error: Can't resolve './src' in '/Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp'","details":"resolve './src' in '/Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp'\n  using description file: /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/package.json (relative path: .)\n    Field 'browser' doesn't contain a valid alias configuration\n    using description file: /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/package.json (relative path: ./src)\n      no extension\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/src is not a file\n      .js\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/src.js doesn't exist\n      .json\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/src.json doesn't exist\n      .wasm\n        Field 'browser' doesn't contain a valid alias configuration\n        /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/src.wasm doesn't exist\n      as directory\n        existing directory /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/src\n          using description file: /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/package.json (relative path: ./src)\n            using path: /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/src/index\n              using description file: /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/package.json (relative path: ./src/index)\n                no extension\n                  Field 'browser' doesn't contain a valid alias configuration\n                  /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/src/index doesn't exist\n                .js\n                  Field 'browser' doesn't contain a valid alias configuration\n                  /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/src/index.js doesn't exist\n                .json\n                  Field 'browser' doesn't contain a valid alias configuration\n                  /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/src/index.json doesn't exist\n                .wasm\n                  Field 'browser' doesn't contain a valid alias configuration\n                  /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/src/index.wasm doesn't exist","stack":"ModuleNotFoundError: Module not found: Error: Can't resolve './src' in '/Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp'\n    at /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/webpack/lib/Compilation.js:2094:28\n    at /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/webpack/lib/NormalModuleFactory.js:895:13\n    at eval (eval at create (/Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/tapable/lib/HookCodeFactory.js:33:10), <anonymous>:10:1)\n    at /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/webpack/lib/NormalModuleFactory.js:332:22\n    at eval (eval at create (/Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/tapable/lib/HookCodeFactory.js:33:10), <anonymous>:9:1)\n    at /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/webpack/lib/NormalModuleFactory.js:509:22\n    at /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/webpack/lib/NormalModuleFactory.js:151:11\n    at /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/webpack/lib/NormalModuleFactory.js:767:25\n    at /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/webpack/lib/NormalModuleFactory.js:979:8\n    at /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/webpack/lib/NormalModuleFactory.js:1109:5"}],"errorsCount":1,"warnings":[{"message":"configuration\nThe 'mode' option has not been set, webpack will fallback to 'production' for this value.\nSet 'mode' option to 'development' or 'production' to enable defaults for each environment.\nYou can also set it to 'none' to disable any default behavior. Learn more: https://webpack.js.org/configuration/mode/","stack":"NoModeWarning: configuration\nThe 'mode' option has not been set, webpack will fallback to 'production' for this value.\nSet 'mode' option to 'development' or 'production' to enable defaults for each environment.\nYou can also set it to 'none' to disable any default behavior. Learn more: https://webpack.js.org/configuration/mode/\n    at /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/webpack/lib/WarnNoModeSetPlugin.js:20:30\n    at Hook.eval [as call] (eval at create (/Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/tapable/lib/HookCodeFactory.js:19:10), <anonymous>:21:1)\n    at Hook.CALL_DELEGATE [as _call] (/Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/tapable/lib/Hook.js:14:14)\n    at Compiler.newCompilation (/Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/webpack/lib/Compiler.js:1254:30)\n    at /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/webpack/lib/Compiler.js:1299:29\n    at Hook.eval [as callAsync] (eval at create (/Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/tapable/lib/HookCodeFactory.js:33:10), <anonymous>:6:1)\n    at Hook.CALL_ASYNC_DELEGATE [as _callAsync] (/Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/tapable/lib/Hook.js:18:14)\n    at Compiler.compile (/Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/webpack/lib/Compiler.js:1294:28)\n    at /Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/webpack/lib/Compiler.js:573:12\n    at Compiler.readRecords (/Users/lukaszjablonski/Desktop/PROJEKTY/AngularBlogApp/node_modules/webpack/lib/Compiler.js:1103:5)"}],"warningsCount":1,"children":[]}
